requirements.txt
discord.py[voice]==2.2.2
python-a2s